+++ CHANGELOG.md
@@ -90,6 +90,7 @@ Note the following more specific changes to these related classes:
   the promises returned by the service.
 * Added alpha version of `DashContainer` for building dynamic, draggable dashboard-style layouts.
   Please note: the API for this component is subject to change - use at your own risk!
+* `Select` now allows the use of objects as values.
 * Added a new `xhEnableImpersonation` config to enable or disable the ability of Hoist Admins to
   impersonate other users. Note that this defaults to `false`. Apps will need to set this config to
   continue using impersonation. (Note that an update to hoist-core 6.4+ is required for this config